pyproject.toml
[project]
name = "did"
version = "0.1.0"
description = "A CLI tool to anonymize Markdown, plain text, TeX, and BibTeX files with spaCy-based entity detection and automatic YAML configuration"
authors = [{ name = "evid 210122431+evidlabel@users.noreply.github.com" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "pyyaml>=6.0.0,<7",
    "flashtext~=2.7",
    "spacy>=3.8.2,<4",
    "en-core-web-md",
    "en-core-web-lg",
    "da_core_news_sm",  # Added for Danish language support
    "da_core_news_md",  # Added for Danish language support
    "click>=8.1.0,<9",
    "rapidfuzz>=3.10.0,<4",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "bibtexparser",
    "rich>=13.3.0,<14",
]

[project.scripts]
did = "did.cli:main"

[dependency-groups]
dev = ["pytest>=8.4.0,<9"]

[tool.uv]

[tool.uv.sources]
en-core-web-md = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl" }
en-core-web-lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl" }
da_core_news_sm = { url = "https://github.com/explosion/spacy-models/releases/download/da_core_news_sm-3.8.0/da_core_news_sm-3.8.0-py3-none-any.whl" }  # Added for Danish support
da_core_news_md = { url = "https://github.com/explosion/spacy-models/releases/download/da_core_news_md-3.8.0/da_core_news_md-3.8.0-py3-none-any.whl" }  # Added for Danish support

[tool.hatch.build.targets.sdist]
include = ["src/did"]

[tool.hatch.build.targets.wheel]
include = ["src/did"]

[tool.hatch.build.targets.wheel.sources]
"src/did" = "did"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
